+++ middleware/cache/config.go
@@ -4,7 +4,6 @@ import (
 	"time"
 
 	"github.com/gofiber/fiber/v3"
-	"github.com/gofiber/utils/v2"
 )
 
 // Config defines the config for middleware.
@@ -27,7 +26,7 @@ type Config struct {
 	// Key allows you to generate custom keys, by default c.Path() is used
 	//
 	// Default: func(c fiber.Ctx) string {
-	//   return utils.CopyString(c.Path())
+	//   return c.App().CopyString(c.Path())
 	// }
 	KeyGenerator func(fiber.Ctx) string
 
@@ -82,7 +81,7 @@ var ConfigDefault = Config{
 	CacheControl:     false,
 	CacheInvalidator: nil,
 	KeyGenerator: func(c fiber.Ctx) string {
-		return utils.CopyString(c.Path())
+		return c.App().CopyString(c.Path())
 	},
 	ExpirationGenerator:  nil,
 	StoreResponseHeaders: false,